moved hook to the app.ts, added new design for notification
