+++ packages/frontend/.env.template
@@ -15,5 +15,5 @@ PORT=3000
 REACT_APP_VERSION=$npm_package_version
 
 # Github repo info
-GITHUB_REPO_OWNER=commons-stack
-GITHUB_REPO_NAME=praise
+REACT_APP_GITHUB_REPO_OWNER=commons-stack
+REACT_APP_GITHUB_REPO_NAME=praise
